requirements.txt
langchain
faiss-cpu
sentence-transformers
pypdf
ipywidgets